05092023 - Component 1.2: Circular Buttons But Failed Output
I made changes to the buttons regarding it''s geometry, where instead of the earlier recntagular shape, I used the pillow module to use canvas. Using canvas, I was able to find a way to form the circular images in the wireframes, but is internally a square hitbox hidden by the background color. However, I was unable to fix the problem the font tuple proble, where I may consider to just use the normal tkinter module instead to avoid double usage errors.