18052023 - Component 3.2: Middle Information
This is the first iteration of the help and info window component, where it will have the main text in the middle that teaches how to use the program and also the functionalities such as ability to export, etc. Currently it is empty in the middle as the text is refined furthermore.