12062023 - Component X: Added Boundary & Changed Theme/Randomizer
Through stakeholder consultation on the assessment, other aspects were added such as boundary warnings. A constant was set for the ranges of the two randomizing numbers, where their max value and min value was set to max range * max range and 0.

Theme was altered by using monochromatic hues of blues, where accents are used throughout rather than a plain black and white, as request by some stakeholders.

Subtraction Alogirthm was changed to cater for those that have not yet learnt about negative integers, and therefore the minimum value is set to a 0.